simplified file structure for app
