--- app/src/components/Content/index.jsx
@@ -1,10 +1,0 @@
-import { BrowserRouter } from 'react-router-dom'
-function Content(){
-    return(<>
-
-
-    
-    </>);
-}
-
-export { Content };
